requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
websockets==12.0
redis==5.0.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
openai-whisper==20231117
edge-tts==6.1.9
groq==0.9.0
langdetect==1.0.9
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.1
numpy==1.26.4
soundfile==0.12.1
python-multipart==0.0.9
httpx==0.27.0
apscheduler==3.10.4
asyncpg==0.29.0
aioredis==2.0.1
pytest==8.2.2
pytest-asyncio==0.23.7